+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Postgres/PostgresAsyncLegacyAspPipelineTests.cs
@@ -11,7 +11,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Postgres
 {
     [NetFrameworkTest]
     public class PostgresAsyncLegacyAspPipelineTests : IClassFixture<PostgresBasicMvcFixture>
