+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Postgres/PostgresExecuteScalarTests.cs
@@ -13,7 +13,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Postgres
 {
     [NetFrameworkTest]
     public class PostgresExecuteScalarTests : IClassFixture<PostgresBasicMvcFixture>
