+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Postgres/PostgresIteratorTests.cs
@@ -5,20 +5,21 @@
 using System.Collections.Generic;
 using System.Linq;
 using NewRelic.Agent.IntegrationTestHelpers;
+using NewRelic.Agent.IntegrationTestHelpers.Models;
 using NewRelic.Agent.IntegrationTests.Shared;
 using NewRelic.Agent.UnboundedIntegrationTests.RemoteServiceFixtures;
 using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Postgres
 {
-    [NetCoreTest]
-    public class PostgresIteratorCoreTests : IClassFixture<PostgresBasicMvcCoreFixture>
+    [NetFrameworkTest]
+    public class PostgresIteratorTests : IClassFixture<PostgresBasicMvcFixture>
     {
-        private readonly PostgresBasicMvcCoreFixture _fixture;
+        private readonly PostgresBasicMvcFixture _fixture;
 
-        public PostgresIteratorCoreTests(PostgresBasicMvcCoreFixture fixture, ITestOutputHelper output)
+        public PostgresIteratorTests(PostgresBasicMvcFixture fixture, ITestOutputHelper output)
         {
             _fixture = fixture;
             _fixture.TestLogger = output;
@@ -41,20 +42,17 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
         [Fact]
         public void Test()
         {
-            var expectedSyncTransactionName = "WebTransaction/MVC/Postgres/PostgresIteratorTest";
-            var expectedAsyncTransactionName = "WebTransaction/MVC/Postgres/PostgresAsyncIteratorTest";
-
+            var expectedSyncTransactionName = "WebTransaction/MVC/PostgresController/PostgresIteratorTest";
+            var expectedAsyncTransactionName = "WebTransaction/MVC/PostgresController/PostgresAsyncIteratorTest";
             var expectedDatastoreCallCount = 2;
 
-            // These values are dictated by the queries that are being run as part of this test.
-            // There are two application endpoints being exercised by the test, each of which runs a query that returns a single row.
-            // The typical pattern in this case is for there to be a call to Read(), followed by a call to NextResult(), followed by a final call to
-            // Read() which returns false to exit the loop.  Each of these roll up to Iterate for a total of 3 for each endpoint.
-            // For the time being, these are commented out due to inconsistent results from one system to another.
-
-            //var expectedSyncIterationCount = 3;
-            //var expectedAsyncIterationCount = 3;
-            //var expectedIterationCount = expectedSyncIterationCount + expectedAsyncIterationCount;
+            //These values are dictated by the queries that are being run as part of this test.
+            //There are two application endpoints being exercised by the test, each of which runs a query that returns a single row.
+            //The typical pattern in this case is for there to be a call to Read(), followed by a call to NextResult(), followed by a final call to
+            //Read() which returns false to exit the loop.  Each of these roll up to Iterate for a total of 3 for each endpoint.
+            var expectedSyncIterationCount = 3;
+            var expectedAsyncIterationCount = 3;
+            var expectedIterationCount = expectedSyncIterationCount + expectedAsyncIterationCount;
 
             var expectedMetrics = new List<Assertions.ExpectedMetric>
             {
@@ -69,9 +67,9 @@ namespace NewRelic.Agent.UnboundedIntegrationTests
                 new Assertions.ExpectedMetric { metricName = @"Datastore/statement/Postgres/teammembers/select", callCount = 1, metricScope = expectedAsyncTransactionName},
 
                 // NpgsqlDataReader methods Read/ReadAsync and NextResult/NextResultAsync result in Iterate metrics.
-                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate" },
-                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", metricScope = expectedSyncTransactionName},
-                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", metricScope = expectedAsyncTransactionName}
+                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", callCount = expectedIterationCount },
+                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", callCount = expectedSyncIterationCount, metricScope = expectedSyncTransactionName},
+                new Assertions.ExpectedMetric { metricName = @"DotNet/DatabaseResult/Iterate", callCount = expectedAsyncIterationCount, metricScope = expectedAsyncTransactionName}
             };
             var unexpectedMetrics = new List<Assertions.ExpectedMetric>
             {
