+++ tests/Agent/IntegrationTests/UnboundedIntegrationTests/Postgres/PostgresStoredProcedureTests.cs
@@ -12,7 +12,7 @@ using NewRelic.Testing.Assertions;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace NewRelic.Agent.UnboundedIntegrationTests
+namespace NewRelic.Agent.UnboundedIntegrationTests.Postgres
 {
     [NetFrameworkTest]
     public class PostgresStoredProcedureTests : IClassFixture<PostgresBasicMvcFixture>
